+++ src/Servers/Linux.php
@@ -14,9 +14,13 @@ use Innmind\Server\Status\{
     Server\Disk,
     Server\Disk\UnixDisk,
 };
+use Innmind\Server\Control\Server as Control;
 use Innmind\TimeContinuum\Clock;
 use Innmind\Url\Path;
-use Innmind\Immutable\Maybe;
+use Innmind\Immutable\{
+    Maybe,
+    Map,
+};
 
 final class Linux implements Server
 {
@@ -26,13 +30,16 @@ final class Linux implements Server
     private LoadAverageFacade $loadAverage;
     private UnixDisk $disk;
 
-    public function __construct(Clock $clock)
+    /**
+     * @param Map<non-empty-string, string> $environment
+     */
+    public function __construct(Clock $clock, Control $control, Map $environment)
     {
-        $this->cpu = new CpuFacade;
-        $this->memory = new MemoryFacade;
-        $this->processes = new UnixProcesses($clock);
+        $this->cpu = new CpuFacade($control->processes());
+        $this->memory = new MemoryFacade($control->processes());
+        $this->processes = new UnixProcesses($clock, $control->processes());
         $this->loadAverage = new LoadAverageFacade;
-        $this->disk = new UnixDisk;
+        $this->disk = new UnixDisk($control->processes());
     }
 
     public function cpu(): Maybe