+++ tests/Facade/Cpu/OSXFacadeTest.php
@@ -7,17 +7,32 @@ use Innmind\Server\Status\{
     Facade\Cpu\OSXFacade,
     Server\Cpu,
 };
+use Innmind\Server\Control\ServerFactory as Control;
+use Innmind\TimeContinuum\Earth\Clock;
+use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Stream\Streams;
 use PHPUnit\Framework\TestCase;
 
 class OSXFacadeTest extends TestCase
 {
+    private $server;
+
+    public function setUp(): void
+    {
+        $this->server = Control::build(
+            new Clock,
+            Streams::fromAmbientAuthority(),
+            new Usleep,
+        );
+    }
+
     public function testInterface()
     {
         if (\PHP_OS !== 'Darwin') {
             $this->markTestSkipped();
         }
 
-        $facade = new OSXFacade;
+        $facade = new OSXFacade($this->server->processes());
 
         $this->assertInstanceOf(Cpu::class, $facade()->match(
             static fn($cpu) => $cpu,
@@ -31,7 +46,7 @@ class OSXFacadeTest extends TestCase
             $this->markTestSkipped();
         }
 
-        $this->assertNull((new OSXFacade)()->match(
+        $this->assertNull((new OSXFacade($this->server->processes()))()->match(
             static fn($cpu) => $cpu,
             static fn() => null,
         ));