+++ tests/Facade/Memory/OSXFacadeTest.php
@@ -7,17 +7,36 @@ use Innmind\Server\Status\{
     Facade\Memory\OSXFacade,
     Server\Memory,
 };
+use Innmind\Server\Control\ServerFactory as Control;
+use Innmind\TimeContinuum\Earth\Clock;
+use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Stream\Streams;
+use Innmind\Immutable\Map;
 use PHPUnit\Framework\TestCase;
 
 class OSXFacadeTest extends TestCase
 {
+    private $server;
+
+    public function setUp(): void
+    {
+        $this->server = Control::build(
+            new Clock,
+            Streams::fromAmbientAuthority(),
+            new Usleep,
+        );
+    }
+
     public function testInterface()
     {
         if (\PHP_OS !== 'Darwin') {
             $this->markTestSkipped();
         }
 
-        $facade = new OSXFacade;
+        $facade = new OSXFacade($this->server->processes(), Map::of([
+            'PATH',
+            $_SERVER['PATH'],
+        ]));
 
         $this->assertInstanceOf(Memory::class, $facade()->match(
             static fn($memory) => $memory,
@@ -31,7 +50,12 @@ class OSXFacadeTest extends TestCase
             $this->markTestSkipped();
         }
 
-        $this->assertNull((new OSXFacade)()->match(
+        $facade = new OSXFacade($this->server->processes(), Map::of([
+            'PATH',
+            $_SERVER['PATH'],
+        ]));
+
+        $this->assertNull($facade()->match(
             static fn($memory) => $memory,
             static fn() => null,
         ));