+++ tests/Server/Processes/UnixProcessesTest.php
@@ -10,18 +10,35 @@ use Innmind\Server\Status\{
     Server\Process\Pid,
     Clock\PointInTime\Delay,
 };
+use Innmind\Server\Control\ServerFactory as Control;
 use Innmind\TimeContinuum\{
     Clock as ClockInterface,
     Earth\Clock,
 };
+use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Stream\Streams;
 use Innmind\Immutable\Set;
 use PHPUnit\Framework\TestCase;
 
 class UnixProcessesTest extends TestCase
 {
+    private $processes;
+
+    public function setUp(): void
+    {
+        $this->processes = new UnixProcesses(
+            new Clock,
+            Control::build(
+                new Clock,
+                Streams::fromAmbientAuthority(),
+                new Usleep,
+            )->processes(),
+        );
+    }
+
     public function testInterface()
     {
-        $this->assertInstanceOf(Processes::class, new UnixProcesses(new Clock));
+        $this->assertInstanceOf(Processes::class, $this->processes);
     }
 
     public function testAll()
@@ -30,7 +47,7 @@ class UnixProcessesTest extends TestCase
             $this->markTestSkipped();
         }
 
-        $all = (new UnixProcesses(new Clock))->all();
+        $all = $this->processes->all();
 
         $this->assertInstanceOf(Set::class, $all);
         $this->assertNotEmpty($all);
@@ -53,7 +70,8 @@ class UnixProcessesTest extends TestCase
             $this->markTestSkipped();
         }
 
-        $process = (new UnixProcesses(new Clock))
+        $process = $this
+            ->processes
             ->get(new Pid(1))
             ->match(
                 static fn($process) => $process,
@@ -67,7 +85,8 @@ class UnixProcessesTest extends TestCase
     public function testReturnNothingWhenProcessDoesntExist()
     {
         $this->assertNull(
-            (new UnixProcesses(new Clock))
+            $this
+                ->processes
                 ->get(new Pid(42424))
                 ->match(
                     static fn($process) => $process,
@@ -82,7 +101,8 @@ class UnixProcessesTest extends TestCase
             $this->markTestSkipped();
         }
 
-        $process = (new UnixProcesses(new Clock))
+        $process = $this
+            ->processes
             ->all()
             ->find(static fn($process) => $process->pid()->is(1))
             ->match(